requirements.txt
flask
werkzeug
pillow
torch
transformers
google-generativeai
python-dotenv
